npm install express-session uuid
express-session is a library that manages user sessions for you. It provides the functionality to save session data in a storage that you choose. You can provide ready-made stores using different libraries like connect-redis. By default, it uses an in-memory store. To make the session persistent throughout recurring requests an identifier is saved inside a cookie.

First of all, we need to install the new dependencies. Additionally to express-session we will use uuid to create the above-mentioned session identifier.